Add style tags inside React components
Not a good solution; only a working solution. Applies styles as expected to React components within the custom element Downside is that this is not the normal way to implement React component styling so each React component essentially has knowledge of being integrated within a custom element as a MFE. to-string-loader is probably also not necessary -- could use css-loader with exportType to get the same output.